Moved consensus and dedup to misc package
