+++ misc/consensus.py
